backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.12.5
python-multipart==0.0.6
sentence-transformers==2.7.0
transformers==4.36.2
peft==0.7.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
numpy==1.24.3
faiss-cpu==1.7.4
pdfplumber==0.10.3
scikit-learn==1.3.2
PyYAML==6.0.3
python-dotenv==1.2.1
accelerate==0.25.0
bitsandbytes==0.41.3